Make height/weight decimal columns

+++ priv/repo/migrations/20170815072107_create_pokemons.exs
@@ -12,8 +12,8 @@ defmodule Pokedex.Repo.Migrations.CreatePokemons do
       add :special_attack, :integer, null: false
       add :special_defense, :integer, null: false
       add :speed, :integer, null: false
-      add :height, :float, precision: 2, null: false
-      add :weight, :float, precision: 2, null: false
+      add :height, :decimal, null: false, precision: 2
+      add :weight, :decimal, null: false , precision: 2
 
       timestamps()
     end
